Optimization - This may be reverted. Dynamic array size change from 128 to 512.
Added min_max vals to dynamic array to try and improve the complexcity of contains.

+++ Project_DeDup/dedup_script.sh
@@ -34,9 +34,13 @@ function run_dedup {
 }
 
 rm -rf time.csv
-echo "#`date`" >> runs_old.csv
-cat runs.csv >> runs_old.csv
-rm -rf runs.csv
+if [ ! -f "runs.csv"]
+then
+	cat runs.csv >> runs_old.csv
+	rm -rf runs.csv
+fi
+
+echo "#`date`" > runs.csv
 
 for FILE in $( ls -Sr | grep ".csv" | cut -f 1 -d . | grep -v "runs" ); do
     rm -rf ${FILE}
